requirements.txt
google-generativeai
requests
python-dotenv
tqdm
rich
textstat